+++ packages/connect-wallet-modal/src/components/WalletsModalForEth/WalletsModalForEth.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { helpers } from '@reef-knot/web3-react';
-import { walletDataList } from '@reef-knot/core-react';
+import { useReefKnotContext } from '@reef-knot/core-react';
+import { WalletConnectLegacyConnector } from 'wagmi/connectors/walletConnectLegacy';
+import { InjectedConnector } from 'wagmi/connectors/injected';
+import { WalletAdapterData } from '@reef-knot/types';
 import {
-  ConnectAmbire,
-  ConnectBlockchaincom,
   ConnectBraveWallet,
   ConnectCoin98,
   ConnectCoinbase,
@@ -15,10 +16,8 @@ import {
   ConnectMetamask,
   ConnectOperaWallet,
   ConnectTrust,
-  ConnectWalletConnect,
+  ConnectWC,
   ConnectXdefi,
-  ConnectZenGo,
-  ConnectZerion,
 } from '../../connectButtons';
 import { ButtonsCommonProps, WalletsModal } from '../WalletsModal';
 import { WalletsModalForEthProps } from './types';
@@ -26,22 +25,18 @@ import { WALLET_IDS, WalletId } from '../../constants';
 
 const walletsButtons: { [K in WalletId | string]: React.ComponentType } = {
   Injected: ConnectInjected,
+  WalletConnectV1: ConnectWC,
   [WALLET_IDS.METAMASK]: ConnectMetamask,
-  [WALLET_IDS.WALLET_CONNECT]: ConnectWalletConnect,
   [WALLET_IDS.LEDGER]: ConnectLedger,
   [WALLET_IDS.COINBASE]: ConnectCoinbase,
   [WALLET_IDS.TRUST]: ConnectTrust,
   [WALLET_IDS.IM_TOKEN]: ConnectImToken,
   [WALLET_IDS.COIN98]: ConnectCoin98,
   [WALLET_IDS.MATH_WALLET]: ConnectMathWallet,
-  [WALLET_IDS.AMBIRE]: ConnectAmbire,
-  [WALLET_IDS.BLOCKCHAINCOM]: ConnectBlockchaincom,
-  [WALLET_IDS.ZENGO]: ConnectZenGo,
   [WALLET_IDS.BRAVE]: ConnectBraveWallet,
   [WALLET_IDS.OPERA]: ConnectOperaWallet,
   [WALLET_IDS.GAMESTOP]: ConnectGamestop,
   [WALLET_IDS.XDEFI]: ConnectXdefi,
-  [WALLET_IDS.ZERION]: ConnectZerion,
 };
 
 function getWalletButton(
@@ -71,19 +66,24 @@ function addWalletTo(
 
 function getWalletsButtons(
   commonProps: ButtonsCommonProps,
+  walletDataList: WalletAdapterData[] = [],
   hiddenWallets: WalletsModalForEthProps['hiddenWallets'] = [],
 ) {
-  let wallets: WalletId[] = [
-    WALLET_IDS.METAMASK,
-    WALLET_IDS.WALLET_CONNECT,
+  let wallets: WalletId[] = [WALLET_IDS.METAMASK];
+
+  // Adding wallets using a new wallet adapter API
+  // TODO: migrate all wallets to use wallet adapter API
+  walletDataList.forEach((walletData) => {
+    const { walletId, detector } = walletData;
+    addWalletTo(wallets, walletId, !!detector?.());
+  });
+
+  wallets = [
+    ...wallets,
     WALLET_IDS.LEDGER,
     WALLET_IDS.COINBASE,
     WALLET_IDS.TRUST,
     WALLET_IDS.IM_TOKEN,
-    WALLET_IDS.AMBIRE,
-    WALLET_IDS.BLOCKCHAINCOM,
-    WALLET_IDS.ZENGO,
-    WALLET_IDS.ZERION,
   ];
 
   // Deprecated way of adding wallets with additional detection
@@ -94,26 +94,8 @@ function getWalletsButtons(
   addWalletTo(wallets, WALLET_IDS.GAMESTOP, helpers.isGamestopProvider());
   addWalletTo(wallets, WALLET_IDS.XDEFI, helpers.isXdefiProvider());
 
-  // Adding wallets using a new wallet adapter API
-  // TODO: migrate all wallets to use wallet adapter API
-  walletDataList.forEach((walletData) => {
-    const { walletId, detector } = walletData;
-    addWalletTo(wallets, walletId, detector());
-  });
-
   // Filtering wallets marked as hidden
   wallets = wallets.filter((wallet) => !hiddenWallets.includes(wallet));
-
-  // Move OKX wallet to the second place
-  // TODO: add better wallets ordering when all wallets migrated to wallet adapter API
-  const okxWalletId = 'okx';
-  const okxWalletIndex = wallets.indexOf(okxWalletId);
-  const metamaskIndex = wallets.indexOf(WALLET_IDS.METAMASK);
-  // If metamask is there and okx was initially placed more than one step away from metamask
-  if (metamaskIndex > -1 && okxWalletIndex > metamaskIndex + 1) {
-    wallets.splice(okxWalletIndex, 1); // remove okx from its place
-    wallets.splice(metamaskIndex + 1, 0, okxWalletId); // place okx right after metamask
-  }
 
   return wallets.map((walletId) => {
     // Handle new wallet adapters
@@ -121,7 +103,14 @@ function getWalletsButtons(
       (data) => data.walletId === walletId,
     );
     if (walletData) {
-      return getWalletButton('Injected', walletId, {
+      let connectorId = '';
+      if (walletData.connector instanceof WalletConnectLegacyConnector) {
+        connectorId = 'WalletConnectV1';
+      }
+      if (walletData.connector instanceof InjectedConnector) {
+        connectorId = 'Injected';
+      }
+      return getWalletButton(connectorId, walletId, {
         ...commonProps,
         ...walletData,
       });
@@ -134,10 +123,12 @@ function getWalletsButtons(
 export function WalletsModalForEth(
   props: WalletsModalForEthProps,
 ): JSX.Element {
+  const { walletDataList } = useReefKnotContext();
+
   return (
     <WalletsModal {...props}>
       {(commonProps: ButtonsCommonProps) =>
-        getWalletsButtons(commonProps, props.hiddenWallets)
+        getWalletsButtons(commonProps, walletDataList, props.hiddenWallets)
       }
     </WalletsModal>
   );