+++ packages/connect-wallet-modal/src/connectButtons/types.ts
@@ -10,3 +10,4 @@ export type ConnectButtonProps = {
 export type ConnectWalletProps = ButtonsCommonProps & ButtonProps;
 
 export type ConnectInjectedProps = WalletAdapterData & ConnectWalletProps;
+export type ConnectWCProps = WalletAdapterData & ConnectWalletProps;