+++ packages/connect-wallet-modal/src/constants/wallets.ts
@@ -1,20 +1,15 @@
 export const WALLET_IDS = {
   METAMASK: 'Metamask',
-  WALLET_CONNECT: 'WalletConnect',
   LEDGER: 'Ledger',
   COINBASE: 'Coinbase',
   TRUST: 'Trust',
   IM_TOKEN: 'ImToken',
   COIN98: 'Coin98',
   MATH_WALLET: 'MathWallet',
-  AMBIRE: 'Ambire',
-  BLOCKCHAINCOM: 'Blockchain.com Wallet',
-  ZENGO: 'ZenGo',
   BRAVE: 'Brave Wallet',
   OPERA: 'Opera Wallet',
   GAMESTOP: 'Gamestop',
   XDEFI: 'Xdefi',
-  ZERION: 'Zerion',
 } as const;
 
 export type WalletId = typeof WALLET_IDS[keyof typeof WALLET_IDS] | string;