+++ packages/core-react/src/hooks/useReefKnotContext.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { ReefKnotContext, ReefKnotContextValue } from '../context';
+
+export const useReefKnotContext = (): ReefKnotContextValue => {
+  return useContext(ReefKnotContext);
+};