+++ packages/core-react/src/index.ts
@@ -1,2 +1,3 @@
 export * from './walletData';
-export * from './utils';
+export * from './context';
+export * from './hooks';