+++ packages/wallets-helpers/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './isValidHttpUrl';
+export * from './user-agents';