+++ packages/web3-react/src/helpers/providerDetectors.ts
@@ -60,7 +60,6 @@ export const isMathWalletProvider = (): boolean => {
 
 export const isImTokenProvider = (): boolean => {
   try {
-    // @ts-expect-error wagmi redeclares window.ethereum type
     return !!window.ethereum?.isImToken;
   } catch (error) {
     return false;
@@ -139,7 +138,6 @@ export const isGamestopInstalled = (): boolean => {
 
 export const isGamestopProvider = (): boolean => {
   try {
-    // @ts-expect-error wagmi redeclares window.ethereum type
     return !!window.ethereum?.isGamestop;
   } catch (error) {
     return false;
@@ -178,7 +176,6 @@ export const isXdefiProvider = (): boolean => {
 
 export const isZerionProvider = (): boolean => {
   try {
-    // @ts-expect-error wagmi redeclares window.ethereum type
     return !!window.ethereum?.isZerion;
   } catch (error) {
     return false;