+++ packages/wallets/exodus/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @reef-knot/wallet-adapter-exodus
 
+## 1.1.0
+
+### Minor Changes
+
+- faa8815: Update wagmi to v0.12
+
+### Patch Changes
+
+- Updated dependencies [faa8815]
+  - @reef-knot/types@1.1.0
+
 ## 1.0.1
 
 ### Patch Changes